requirements.txt
ttkthemes==3.2.2
reportlab==4.0.4
pandas==2.1.1
openpyxl==3.1.2
pillow==10.0.1
python-dateutil==2.8.2
pyinstaller==6.1.0 